database_support/requirements.txt
openai==1.3.0
python-dotenv==1.0.0
rich==13.7.0
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1 